requirements.txt
pytesseract~=0.3.10
Pillow~=9.4.0
pdf2image~=1.16.3
opencv-python~=4.7.0.72
numpy~=1.24.1
pandas~=1.5.2